Bugfix/91 renaming button bar when no objective selected throws error (#96)
* The Start State is now the selected Objective from the beginning
* Fixes Bugs #91 #94  #95 
* Adds links to error / warning messages in the dependency modeler
---------

+++ app/lib/objectivemodeler/buttonbar/OmButtonBar.js
@@ -80,10 +80,12 @@ export default function OmButtonBar(canvas, eventBus, omModeler) {
     deleteObjectiveButton.title = 'Delete Current Objective';
     deleteObjectiveButton.addEventListener('click', () => {
         var objectiveToDelete = selectObjectiveComponent.value;
-        var shouldDelete = eventBus.fire(ObjectiveEvents.OBJECTIVE_DELETION_REQUESTED, { objective: objectiveToDelete});
-        if (shouldDelete !== false) {
-            // Deletion was not rejected and not handled somewhere else; should not happen when mediator is involved
-            omModeler.deleteObjective(objectiveToDelete);
+        if( objectiveToDelete.id !== 'StartBoard') {
+            var shouldDelete = eventBus.fire(ObjectiveEvents.OBJECTIVE_DELETION_REQUESTED, {objective: objectiveToDelete});
+            if (shouldDelete !== false) {
+                // Deletion was not rejected and not handled somewhere else; should not happen when mediator is involved
+                omModeler.deleteObjective(objectiveToDelete);
+            }
         }
         selectObjectiveComponent.showValue(omModeler.getCurrentObjective());
     });
